+++ src/ConsulSharp/V1/Transaction/Models/ITransactionOperation.cs
@@ -1,0 +1,13 @@
+﻿namespace ConsulSharp.V1.Transaction.Models
+{
+    /// <summary>
+    /// The Transaction Operation interface.
+    /// </summary>
+    public interface ITransactionOperation
+    {
+        /// <summary>
+        /// The operation type.
+        /// </summary>
+        string OperationType { get; }
+    }
+}