+++ src/ConsulSharp/V1/Transaction/Models/ITransactionResult.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace ConsulSharp.V1.Transaction.Models
+{
+    /// <summary>
+    /// The transaction result interface.
+    /// </summary>
+    [JsonConverter(typeof(TransactionResultJsonConverter))]
+    public interface ITransactionResult
+    {
+        /// <summary>
+        /// The operation type.
+        /// </summary>
+        string OperationType { get; }
+    }
+}