+++ src/ConsulSharp/V1/Transaction/Models/TransactionStatus.cs
@@ -1,0 +1,12 @@
+﻿namespace ConsulSharp.V1.Transaction.Models
+{
+    /// <summary>
+    /// The Transaction Status enum.
+    /// </summary>
+    public enum TransactionStatus
+    {
+        None = 0,
+        Success = 1,
+        Rollback = 2,
+    }
+}