+++ CHANGELOG.md
@@ -7,3 +7,4 @@
  * All KeyValue APIs - Write, Read, Delete Value.
  * All Snapshot APIs - Generate and Restore.
  * All Status APIs - Get leader and Raft peers.
+ * All Transaction APIs - Create with strong verbs and strong operation types for KV and strong results.